[#146] refactor - 한국어 ko로 변경
